Make term_memberships work with direct memberships
If a Membership includes a `legislative_period`, use it, rather than
calculating on term dates.

The `suomi.rb` test is identical to the `eduskunta.rb` test, other than
in which data file it uses. There’s almost certainly a way to share
this, but let’s just get it working for now.