Add basic Party Groupings box
Add a count of how many seats each party group had during this Term.
It’s not ideal, as it simply counts the number of Memberships over the
Term, but it’s a good first step, and lets us style the page better
whilst we work out what the correct calculations should be.