Countries are valid if the data exists
Rather than maintaining a list of country/filename mappings, simply
enable a new country just by dropping it in the `data` directory.

This removes the ability to have multiple URLs to a country, but that’s
not really a necessary feature, and if we want we can bring it back
later in a different way.