Remove Popolo::Data
We’re only calling it from one place, and it’s an outdated concept.
Instead simply instantiate a`GithubFile` for the relevant CSV and
Popolo.

(As this effectively lets us now use the correctly passed legislature
for a country, we need to change the per-country page to only display
for the first one it knows about. Tidying that up is part of another
change.