Remove the last_modified header
We added this in the hope that it would speed up building the site, but
(a) it didn't, as wget doesn't DTRT with it, and (b) we spend the build
up substantially in other ways anyway. All this does now is make so
that we have to do hard-refreshes when we change the layout (rather
than the data), so remove it.