Move JSON download button to avoid confusion over scope
Previously JSON download was presented alongside CSV download
for a legislature’s membership table, despite the two files
containing very different data.

This moves the JSON download link to the top of the term page
(and names it clearly with the number of terms it contains)
and also adds JSON download links to the country page.

Fixes #370.