Correctly limit width of homepage autocomplete dropdown
Fixes a styling bug related to #769.

Also, since menu items are more likely to wrap onto multiple
lines now, we increase the line-height slightly to make them
look better when they do wrap.

And for consistency, we use the same prototype-extending
paradigm for the other autocomplete hack we added in
338648794956f315b7adca356caa8b6b99e7d792.