Better $.sortable() handling for pre-sorted columns
The first click on a pre-sorted column header (eg: "Name") will
reverse the sort order, and the second click will just revert
to the original sort order, rather than attempting to "unsort"
the column.

Pre-sorted columns will also be marked as such in the UI, with
an arrow pointing in the right direction. This should make it
much more obvious what’s going on.