Extract person rendering into a separate class
This handles getting people and their memberships out of the popolo and
restructuring them for the templates.

This code could be much nicer with a few changes to
everypolitician-popolo. In particular it would be nice if that lib
handled filtering by term. Other useful things would be if it always
defined methods so that you could call methods without needing to know
if they existed on the target person.