refactor injection of query vars into URL for clarity
code now more obviously checks for hash (fragment identifier in URL)
and simply adds the query vars on the end of the URL if there isn't
one there.

We don't anticipate there being an anchor tag there any more, but
it's a sneaky gotcha that would break this behaviour should one
ever be added.