restrict hiding to this list only
This is just in set-up of the list: an alternative approach would be to
have the hidden class have `display:none` in the stylesheet. I haven't
done it that way simply because hiding the elements dynamically like this
means any class can be used. Since this is only dynamically controlled,
I've added the prefix `js-` to the hidden class name being used.