extract innerList logic into its own function
makes it a teeny bit clearer what's going on.
I think the function itself is a little clunky because of the problem of
the targetIndex being at (or close to) either end of the list, so there's
some explicit checking for that.

Also it makes it clearer that this code is adding the hidden class to
everything, then removing it from the innerList... would be nicer to
invert this to add the class just to things not in the innerList, which
might be possible for cleanliness.