Split the exhaustive web tests to a separate rake task
Running the entire test suite takes forever because it has to make sure
that every country works as a fully rendered page.

If we split those out to a separate `rake test:web` target, and ensure
that `rake test:all` runs all of them, then we can still run the full
test suite when needed, but the day-to-day run of the Page tests can be
lightning fast.