Update the country route and the templates it uses
We are setting the title temporarily in he layout template  to use the member field of the page object while the other pages are not extracted.
In the future all pages may have a title field that we could use in the main layout template.
We are also setting the country variable if @page.country is available for the same reason.