Tidy up Download class
Now that everypolitician-ruby returns `nil` rather than throwing an
exception if a country doesn't exist, we need to handle that case
better. But even though the resulting code is simpler, it's still too
complex for the initializer, which shouldn't really be doing anything
other than, erm, initialising. So pull it out to separate method.