require ruby 2.3.1
As we're moving to using named parameters everywhere, 2.0.0 isn't really
good enough anymore — we want at least 2.1.0 so we get required parameters.

With libraries we want to stick to the oldest plausible version rather
than forcing users to upgrade their ruby, but with an app like this we can be
a lot more relaxed. So if we're upgrading anyway, we may as well go to
2.3.1