Add a `person_count` to Everypolitician::Country
EveryPolitician::Country only has a person_count on a Legislature.
Let's pretend it also has one on a Country.

We might want to actually _add_ this to `everypolitician-ruby`, but
it's not totally clear year whether it's something that's generally
useful enough, so let's live with it here for a while first, and then
decide.