Switch card test to Fichtinger
Testing the card for Angela Fichtinger rather than Andrea Gessl-Ranftl
lets us ensure that all the cards are correctly filled, as Andrea
doesn't have any 'social' data. Also rename the test to make it clearer
that we're testing the construction of the card, rather than just
getting the correct list of people.