Update the world helper's struct
The url member of the Country struct was renamed to slug to play well with the real Country objects.

The main layout was calling a :url field in the @country returned by the currently displayed page,
but Country objects don't have a url member, they have a :slug instead.

So by making the titles work for world countries, we broke it for EP countries.
After renaming url to slug in world countries, it should work for both types of country.