pass Country object, not slug
If we instantiate a Country object with a slug for a country for which
we don't have data, bad things ensue. Rather than making that the job
of the Page object, have it accept an already-instantiated (and thus
valid) Country object, and make the route responsible for generating
that. This approach should considerably simplify error handling
everywhere and provide a cleaner separation of concerns.