Extend EP-Legislature with a `term` method
So people can call `.country("UK").legislature("Commons").term(65)`

This should almost certainly move to everypolitician-ruby, but it's a
little tricky as a _proper_ version will need to be able to navigate the
boundaries between `everypolitician-ruby` and `everypolitician-popolo`
(e.g. to get the members in that term). I'm not quite sure yet how
that's going to work, so let's explore the issues here first.