Read the DATASOURCE from an environment variable
Previously we were getting the url for the `countries.json` from a
`DATASOURCE` file in the root of the project. This meant that each time
we wanted to update the site we'd need to do a commit updating the
datasource.

The new way we want to do things is allow you to supply an environment
variable or something which contains the url, but default to using
`master` if nothing is specified.