Fix tests after changing where the DATASOURCE comes from
Now we're reading from the `DATASOURCE` environment variable we need to
change the tests to cope with the default url being a GitHub raw url,
rather than a RawGit url. The reason it's now a GitHub raw url is so
that we can point to the master branch, which we can't do with RawGit
because it would cache the results on the CDN indefinitely.