Change GITHUB_ACCESS_TOKEN to be an everypoliticianbot one
Previously this was just used for read operations, so it didn't matter
that this token wasn't owned by everypoliticianbot. Now we want it to
also do write operations, so we will need one owned by the bot.

The command I used to re-generate the secure section of the config was:

    travis encrypt GITHUB_ACCESS_TOKEN=change_to_a_real_access_token

See
https://docs.travis-ci.com/user/environment-variables/#Encrypting-environment-variables
for details on how to create secure environment variables.