Use the redirect method in the country/download route
Since the docs and the country page have different hosts, I made a
meta_redirect method to take an URI as an argument.
The /about.html and docs routes keep their call to docs_redirect,
which calls the meta_redirect with the right URI.

Also, the route is now switching from instantiating Page::Download
to Page::Country, as we are removing the Page::Download from
the codebase to use the HouseDownload instead, and we don't
know which legislature to go to at this point from the route.