Use a .netrc file for Travis' GitHub Access Token
Embedding the access token in the url makes it very easy to accidentally
leak the access token, for example the full url (including token) is
shown after a git push.

This change moves the credentials into a .netrc file, which means they
aren't part of the url, so should be much harder to accidentally leak.

Travis docs on using a .netrc file:
https://docs.travis-ci.com/user/private-dependencies/#API-Token