Add ruby-vnu to Gemfile for HTML validation
Firefox was showing the source code with the heading tags in red.
I discovered that it was due to a missing link-closing tag, so I
added </a> to the main layout heading. But it is much better if we
have an automated way to detect and correct those things. Hence
this addition. The gem uses the same validator than the W3C.