Check for supported version of html-tidy in tests
This change adds a more robust check for a supported version of
html-tidy.

Previously we were only checking that html-tidy was installed, however
this meant that the user might have a very old version of html-tidy
installed, which doesn't support HTML5. This new check now ensures that
the user's version of html-tidy actually supports HTML5.