Remove duplicate before_install section in .travis.yml
When I added the script to install html-tidy I accidentally created a
second, duplicate, `before_install` section. Because of the way YAML
works this silently overwrote the other `before_install` section which
was setting the contents of `.netrc` so that git had some credentials to
use to deploy.

This change combines the two sections into one.