PersonCard: use the 'memberships' attribute of the person
The displayed positions should only be for a particular term, however,
so we need to pass 'term' as well, to select memberships for that term.
This isn't a bad thing, though, because in subsequent commits 'term'
will be useful for getting the legislature, country, proxy image URL
and the top identifiers.