Remove hard-coded SHAs in tests
Having the SHA in the path makes it harder to script updates to the
fixtures and means that the knowledge of the current SHA is duplicated
between `countries.json` and the path.

We do want to track the SHA of the fixtures, but this information is
already in git in the `countries.json` fixture and in the VCR cassettes,
so we don't need to also be tracking it in the directory names in the
fixtures and in the tests themselves.