Update README docs for unstable cabinet memberships
Trying to do this directly in the Popolo classes wasn't working because
the Popolo classes (understandably) don't know anything about the world
outside Popolo.

So to fix this we're going to move the logic for looking up cabinet
memberships into the `TermTable` and `PersonCard` classes for now. Then
once we've found a stable place for the cabinet memberships to live we
can then move more of the logic into the Popolo classes or wherever
makes sense.