Super hacky experiment getting seat count from Wikidata
This is just an experiment to get the seat count for a legislature from
Wikidata. It gets the Wikidata JSON representation for a country and
extracts the seat count from that. It does all of that right inside the
Page class.

There's a bunch of problems with this. It would probably be better to
use a gem to access the Wikidata API. There's code in the Page class
that violates the architecture that says these classes shouldn't know
about the outside world.

And all that's before you get to the fact that we eventually want to do
this in client-side Javascript and not write this code twice. As I said,
it's a super hacky experiment.