Add back the note in the README about how we handle code reloading
In be75028c we removed the note in the README about needing to restart
the webserver process when making changes to `app.rb` because it now
happens automatically thanks to Sinatra::Reloader.

Rather than simply removing this, it's better to state that new reality
explicitly, especially as the new behaviour doesn't seem to be obvious
at runtime, and people's won't necessarily expect that this will happen.