Fetch seat count from Wikidata in client-side JavaScript
We're already querying the API on the server so what's the point of
this, you ask? We'll we'd like to move this to _only_ using client-side
JS so we don't have to write these queries twice.

So soon we'll remove the server-side query. But first we need to work
out what we do with our tests, and eventually our static site
generation.