Don't load seat count on the backend
This removes the Sinatra/Ruby loading of seat count so it's only loaded
using client-side JavaScript. This is where we eventually want to be so
we're not writing code twice to fetch data from Wikidata.

At the moment the integration test is failing because the JS isn't
executed in the test. A subsequent commit will fix this up.