Remove unused Wikidata API stubbing
We're not calling the Wikidata API server-side any more so we're not
using these stubs.

The Wikidata API access is now being done client-side and therefore
tested with capybara-webkit. It might make sense to reimplement this
web request stubbing at some stage to isolate the tests from internet
access. I'm not sure if that's possible with capybara-webkit and it's a
job for another day.