Install capybara-webkit in development and test
We're going to use this gem to scrape pages as part of our static site
generation, after JavaScript has finished running on those pages.

We're putting this in our Gemfile under the test and development groups
so you can use it in a Rakefile during development and by Travis under
test when the static site is built.