Add Rake task to build static site pages that need JavaScript
We want to be able to generate some pages as part of the static site
build after JavaScript has run on that page. This will allow these pages
to be populated with data from JavaScript but appear to progressively
enhance when deployed as part of our static site.

This Rake task currently has a hardcoded list of one page that needs
JavaScript run. This can be expanded pretty easily later. It then uses
capybara-webkit to fetch that page, run JS, and save those pages to
disk. The target files will be in the same location as `wget` uses which
will allow us to overwrite the static files already generated by `wget`
with these JS-enhanced versions.