Rework JS static site code structure
This splits up the generator class as was a problem suspected in this PR
review:

https://github.com/everypolitician/viewer-sinatra/pull/15640#pullrequestreview-90860548

It creates a single class for a Browser which extends the
capybara-webkit browser to include a way to get a page's contents after
jQuery's AJAX calls are done.

This is in a StaticSite module just to namespace it and indicate where
it's used in the application, rather than it's nature. In other words
it's not a _static site browser_ but you could read it that way so maybe
this needs changing?

All the remaining procedural logic is pushed into the Rake task. From
here we could then extract that if we see fit.