Install `tidy` from package on Travis
Previously we had a script that compiled a recent HTML Tidy from source
as the version that ships with Ubuntu Trusty is ancient. This switches
us to using the version that ships in the Debian sid repo which looks to
be the same version we were compiling from source:

https://packages.debian.org/sid/tidy

Removing the custom script and not compiling something from source makes
things a little simpler for us and as a side effect should speed up the
build slightly.