Signpost Wikidata on the homepage
Most of the changes required for https://github.com/everypolitician/everypolitician-docs/issues/43
(Rewrite docs in light of Wikidata-first approach) need to happen in the
`everypolitician-docs` repo, but we also need to highlight the Wikidata
connection better from here too.

We will probably want to take this further / make this clearer here once
we've actually written the rest of those docs, but we need to do
*something* now, as it still looks as if our goal here is to do all this
ourselves.