Remove fragile seat count tests that rely on live Wikidata
Thee tests are incredibly fragile, as they depend on values currently
returned from Wikidata. This whole approach was largely a dead-end
anyway, so it should be safe enough to remove only the tests, and then
we can either remove the code later, or add better tests if we actually
need them.